Rewrite the Coding_Buddy project from scratch without cookiecutter
Started from scretch. Instead of cookiecutter, django-admin startproject
was used. Code was moved from the old app, cleaned, possibly few files
could have been lost in the process and should be copied over.

A Vagrantfile was created but it cannot be considered mature.

The setup stops at `No Facebook app configured` problem, which should
be addressed in a migration or fixture.

TODO:
- add README
- make setup.yml file work for local dev environment setup.
- revise everything